requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
anthropic==0.18.1
openai==1.13.3
groq==0.4.2
toml==0.10.2
python-dateutil==2.9.0
flask-login==0.6.3
bootstrap-flask==2.3.3
gitpython==3.1.42
python-frontmatter==1.0.1
nltk==3.8.1
numpy==1.26.4
joblib==1.3.2
